+++ lib/core.dart
@@ -1,4 +1,3 @@
-export 'core/bottom_modal_sheet.dart';
 export 'core/constants.dart';
 export 'core/database.dart';
 export 'core/dialogs.dart';