+++ lib/modules/radarr/modules/radarr/widgets/appbar_add_movie_action.dart
@@ -7,7 +7,7 @@ class RadarrAppBarAddMoviesAction extends StatelessWidget {
     Widget build(BuildContext context) {
         return LunaIconButton(
             icon: Icons.add,
-            onPressed: () async => RadarrAddMovieRouter().navigateTo(context),
+            onPressed: () async => RadarrAddMovieRouter().navigateTo(context, query: ''),
         );
     }
 }